pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crf-migrate"
version = "0.1.0"
description = "SDTM annotation migration tool for annotated CRF PDFs"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",
    "rapidfuzz>=3.0.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.38.0",
    "pydantic>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["src"]
